쿠폰 할인율 단위를 소수로 변경

+++ src/main/java/site/brainbrain/iqtest/controller/dto/CouponResponse.java
@@ -3,6 +3,6 @@ package site.brainbrain.iqtest.controller.dto;
 import site.brainbrain.iqtest.domain.CouponType;
 
 public record CouponResponse(CouponType couponType,
-                             int discountRate,
+                             double discountRate,
                              boolean isAvailable) {
 }
